modify: 변경된 스타일 및 헤더 숙소 atom 상태 유지 로직 병합
